requirements.txt
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.7
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
safetensors==0.6.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
supervision==0.26.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.227
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
